brifly finish Lab2A

+++ src/utils/utils.go
@@ -1,11 +1,11 @@
 package utils
 
-func GetIntPtr(x int) *int {
+func GetIntPtr(x int64) *int64 {
 	return &x
 }
 
 
-func Min(x, y int) int {
+func Min(x, y int64) int64 {
 	if x < y {
 		return x
 	}
